Show working clock (minutes + seconds)
This includes setting up a JS SetInterval call that triggers a tick event every second.
The event will be injected with the current time and... Currently will do nothing expect for setting this time in the app state.
There's a UI component that listens to this state change.

Now, all I need to do is to make it countdown instead of countup (showing the current time).