Showing working timer with hardcoded 20 minutes
Added a Start button and default 20 minutes timer.
When you press start, the timer will count down.
I still don't have any logic for what happens when the timer reaches 0 so it will probably just continue counting in negatives.
But I'm basically taking the different between now and end time (start time + duration) and format it to be mm:ss.

Added proto-repl as dev dependency for proto-repl to work.

I will need to continue playing with the configuration in order to use proto-repl as figwheel repl.