Can enter duration manually using web form
The countdown doesn't stop though - something's wrong with the paused logic.
Also tried to make proto-repl work with figwheel and clojurescript - failed miserably. It seems that it works for everytone except me.
If it continues like this I'll have to swithc to something more supported and documented like the bloody emacs or even (god-help-please-not) cursive.