Make timer stop at 00:00
Just changed the time-diff function to fallback to 0 if the difference is negative.
It's a bit of business logic inside util logic but I've decided that the code is not a library and is extremely specific. We shouldn't over-generalize it will make code difficultto read with all kinds of if-elses.

Or maybe I'm wrong :)

Also updated the figwheel sidecar version to make sure the hot-reloading works (it actually wasn't before because of old goog.net reload usage).