Added a progress bar that's getting smaller with time
Had to write some ugly code to sync with the time - will probably require some refactoring in the future to make it look bettet.
Currently it work but looks ugly but it's ok since it's only a playground version of the app.

The progress bar is an HTML5 tag with rotation added to make it look vertical. Probably doesn't work on FF and IE browsers since the rotation CSS magic is different for everything. Can be easily fixed later. I hope.