Added the automatic EMOM functionality
The UI is still ugly but now you can enter a list of movements, put in the duration and the timer will figure out which is the active movement for the current minute.
It's a simple math formula based on total number of seconds, seconds left and count of movements.