Added a second progres bar to show the current minute
The current minute is important to know how much time is left for the current exercise.
I've used another simple formula to calculate the current percentage of the current minute based on the total duration and seconds left in a workout.

Also added more styles to visually separate two progress bars.