Created a speak effect
Instead of calling a function from the tick event I've switched to declaring an effect as the result of the tick event. The effect is :speak with a string providing the value that should be pronounced out loud. That allows for a much better separation and cleaner flow of the program.

I will continue reading the docs and refactoring the giant ugly tick event. I can probably do a lot of stuff with interceptors to control the flow and make tick much smaller and more readable.